myTrades remove fixed

+++ app/src/main/java/pt/ipca/smartcanteen/models/adapters/viewHolders/OrdersAdapterRecViewHolder.kt
@@ -91,7 +91,7 @@ class OrdersAdapterRecViewHolder(val progressBar: ProgressBar, val textProgress:
         myOrdersAdapter.adapter = adapter
     }
 
-    fun bindData(identifierText: Int, quantityText: Int, priceText: Int, stateText: String){
+    fun bindData(identifierText: Int, quantityText: Int, priceText: Float, stateText: String){
         identifierTv.text= identifierText.toString()
         quantityTv.text = quantityText.toString()
         priceTv.text= priceText.toString()
